+++ src/Service/TimestreamWrite/src/TimestreamWriteClient.php
@@ -112,6 +112,27 @@ class TimestreamWriteClient extends AbstractApi
         }
 
         switch ($region) {
+            case 'ingest-us-east-1':
+                return [
+                    'endpoint' => 'https://ingest.timestream.ingest-us-east-1.amazonaws.com',
+                    'signRegion' => 'us-east-1',
+                    'signService' => 'timestream',
+                    'signVersions' => ['v4'],
+                ];
+            case 'ingest-us-east-2':
+                return [
+                    'endpoint' => 'https://ingest.timestream.ingest-us-east-2.amazonaws.com',
+                    'signRegion' => 'us-east-2',
+                    'signService' => 'timestream',
+                    'signVersions' => ['v4'],
+                ];
+            case 'ingest-us-west-2':
+                return [
+                    'endpoint' => 'https://ingest.timestream.ingest-us-west-2.amazonaws.com',
+                    'signRegion' => 'us-west-2',
+                    'signService' => 'timestream',
+                    'signVersions' => ['v4'],
+                ];
             case 'ingest-fips-us-east-1':
                 return [
                     'endpoint' => 'https://ingest.timestream-fips.us-east-1.amazonaws.com',